requirements.txt
# Requires memached to be installed.
python-memcached==1.57
Django==1.8.8
Fabric==1.10.2
# If this fails, run sudo apt-get install python-dev libpq-dev
psycopg2==2.6.1
beautifulsoup4==4.3.2
coverage==3.7.1
flup==1.0.2
nltk==2.0.4
ipaddr==2.1.11
newrelic==2.56.0.42
python-whois==0.5.2
uWSGI==2.0.11.1
numpy==1.9.1
csscompressor==0.9.3
python-dateutil==2.4.2
robotexclusionrulesparser==1.6.2
protobuf==2.6.1
protobuf.socketrpc==1.3.2
django-tagging==0.4
# We're no longer using a local version of langid (local folder can be deleted)
langid==1.1.5
wget==2.2
pytz==2015.7
